внесла коррективы кое-где убрала статик

+++ src/core/FileManager.java
@@ -2,24 +2,27 @@ package core;
 
 import ui.Menu;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.*;
 
 public class FileManager {
-    private static String textFileRead;
+    private String textFileRead;
 
-    public static String getTextFileRead() {
+    public String getTextFileRead() {
         return textFileRead;
     }
 
     public void readLine(String path) throws IOException {
         Path file = Path.of(path);
         if (Files.exists(file)) {
-           textFileRead = Files.readString(file);
+            textFileRead = Files.readString(file);
         }
     }
     public void writeLine(String text, String path) throws IOException {
         Path writeFile = Path.of(path);
-        Files.writeString(writeFile, text);
+        if (Files.exists(writeFile)) {
+            Files.writeString(writeFile, text);
+        }
     }
 }
